backend/requirements.txt
flask
flask-cors
flask-socketio
RPi.GPIO
python-dotenv
eventlet
psutil